+++ containers/src/Data/Map.hs
@@ -3,13 +3,6 @@
 {-# LANGUAGE Safe #-}
 #endif
 
-#ifdef __GLASGOW_HASKELL__
-{-# LANGUAGE DataKinds, FlexibleContexts #-}
-#endif
-#if __GLASGOW_HASKELL__ >= 800
-{-# LANGUAGE MonoLocalBinds #-}
-#endif
-
 #include "containers.h"
 
 -----------------------------------------------------------------------------
@@ -67,48 +60,6 @@
 
 module Data.Map
     ( module Data.Map.Lazy
-#ifdef __GLASGOW_HASKELL__
-    , insertWith'
-    , insertWithKey'
-    , insertLookupWithKey'
-    , fold
-    , foldWithKey
-#endif
     ) where
 
 import Data.Map.Lazy
-
-#ifdef __GLASGOW_HASKELL__
-import Utils.Containers.Internal.TypeError
-
--- | This function is being removed and is no longer usable.
--- Use 'Data.Map.Strict.insertWith'.
-insertWith' :: Whoops "Data.Map.insertWith' is gone. Use Data.Map.Strict.insertWith."
-            => (a -> a -> a) -> k -> a -> Map k a -> Map k a
-insertWith' _ _ _ _ = undefined
-
--- | This function is being removed and is no longer usable.
--- Use 'Data.Map.Strict.insertWithKey'.
-insertWithKey' :: Whoops "Data.Map.insertWithKey' is gone. Use Data.Map.Strict.insertWithKey."
-               => (k -> a -> a -> a) -> k -> a -> Map k a -> Map k a
-insertWithKey' _ _ _ _ = undefined
-
--- | This function is being removed and is no longer usable.
--- Use 'Data.Map.Strict.insertLookupWithKey'.
-insertLookupWithKey' :: Whoops "Data.Map.insertLookupWithKey' is gone. Use Data.Map.Strict.insertLookupWithKey."
-                     => (k -> a -> a -> a) -> k -> a -> Map k a
-                     -> (Maybe a, Map k a)
-insertLookupWithKey' _ _ _ _ = undefined
-
--- | This function is being removed and is no longer usable.
--- Use 'Data.Map.Strict.foldr'.
-fold :: Whoops "Data.Map.fold is gone. Use foldr."
-     => (a -> b -> b) -> b -> Map k a -> b
-fold _ _ _ = undefined
-
--- | This function is being removed and is no longer usable.
--- Use 'foldrWithKey'.
-foldWithKey :: Whoops "Data.Map.foldWithKey is gone. Use foldrWithKey."
-            => (k -> a -> b -> b) -> b -> Map k a -> b
-foldWithKey _ _ _ = undefined
-#endif