+++ containers/changelog.md
@@ -1,5 +1,14 @@
 # Changelog for [`containers` package](http://github.com/haskell/containers)
 
+## 0.7
+
+* Drop support for GHC versions before 8.0.
+
+* Remove all `DEPRECATED` functions.
+
+* Bump Cabal version for tests, and use `common` clauses to reduce
+  duplication.
+
 ## 0.6.5.1
 
 ### Bug fixes